+++ src/main/java/way/paved/finger/City.java
@@ -1,11 +1,12 @@
-package wayPavedFinger;
+package way.paved.finger;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @AllArgsConstructor
 @Getter
-public class Monster {
+public class City {
     private String name;
-    private String location;
+    private Location location;
+    private int distance;
 }
