+++ src/main/java/way/paved/finger/CityWorker.java
@@ -1,4 +1,4 @@
-package wayPavedFinger;
+package way.paved.finger;
 
 import lombok.AllArgsConstructor;
