+++ src/main/java/way/paved/finger/Location.java
@@ -1,4 +1,4 @@
-package wayPavedFinger;
+package way.paved.finger;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
