+++ src/main/java/way/paved/finger/Witcher.java
@@ -1,4 +1,4 @@
-package wayPavedFinger;
+package way.paved.finger;
 
 import java.util.ArrayList;
 import java.util.List;
